gauth/requirements.txt
# google client ( for google services )
google-api-python-client==2.154.0
# google auth
google-auth-oauthlib==1.2.1